backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
httpx
tenacity
pypdf
python-docx
pandas
openpyxl
python-pptx
Pillow
redis
